% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splineFunctions.R
\name{alpha}
\alias{alpha}
\title{alpha function}
\usage{
alpha(i, j, knots, knotidx)
}
\arguments{
\item{i}{index i}

\item{j}{index j}

\item{knots}{knot vector}

\item{knotidx}{knot index}
}
\value{
a vector
}
\description{
A function used in calculating the coefficients of a B-spline curve
}
