% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splineFunctions.R
\name{B}
\alias{B}
\title{B function}
\usage{
B(x, i, j, knots)
}
\arguments{
\item{x}{locations at which to evaluate the B-spline}

\item{i}{index i}

\item{j}{index j}

\item{knots}{a knot vector}
}
\value{
a vector of polynomial coefficients
}
\description{
A recursive function used in calculating the coefficients of a B-spline curve
}
