% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseHazST.R
\name{baseHazST}
\alias{baseHazST}
\title{baseHazST function}
\usage{
baseHazST(
  bh1 = NULL,
  survobj,
  t0,
  nbreaks = 5,
  breakmethod = "quantile",
  MLinits = NULL
)
}
\arguments{
\item{bh1}{X}

\item{survobj}{X}

\item{t0}{X}

\item{nbreaks}{X}

\item{breakmethod}{X}

\item{MLinits}{X}
}
\value{
...
}
\description{
A function to
}
