\name{baselinehazard}
\alias{baselinehazard}
\title{baselinehazard function}
\usage{
baselinehazard(x, t = NULL, n = 100, probs = c(0.025, 0.5, 0.975),
  plot = TRUE)
}
\arguments{
  \item{x}{X}

  \item{t}{X}

  \item{n}{X}

  \item{probs}{X}

  \item{plot}{X}
}
\value{
...
}
\description{
A function to
}

