% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_baseline_hazard_multiway.R
\name{baselinehazard_multiWay}
\alias{baselinehazard_multiWay}
\title{baselinehazard_multiWay function}
\usage{
baselinehazard_multiWay(
  x,
  probs = c(0.025, 0.5, 0.975),
  cumulative = FALSE,
  plot = TRUE,
  joint = FALSE,
  xlims = NULL,
  ylims = NULL,
  ...
)
}
\arguments{
\item{x}{X}

\item{probs}{X}

\item{cumulative}{X}

\item{plot}{X}

\item{joint}{X}

\item{xlims}{X}

\item{ylims}{X}

\item{...}{X}
}
\value{
...
}
\description{
A function to
}
