\name{betapriorGauss}
\alias{betapriorGauss}
\title{betapriorGauss function}
\usage{
betapriorGauss(mean, sd)
}
\arguments{
  \item{mean}{the prior mean, a vector of length 1 or more.
  1 implies a common mean.}

  \item{sd}{the prior standard deviation, a vector of
  length 1 or more. 1 implies a common standard deviation.}
}
\value{
an object of class "betapriorGauss"
}
\description{
A function to define Gaussian priors for beta.
}

