% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splineFunctions.R
\name{Bspline.construct}
\alias{Bspline.construct}
\title{Bspline.construct function}
\usage{
Bspline.construct(x, basis)
}
\arguments{
\item{x}{a vector, the data}

\item{basis}{an object created by the getBbasis function}
}
\value{
a basis matrix
}
\description{
A function to construct a B-spline basis matrix for given data and basis coefficients. Used in evaluating the baseline hazard.
}
