\name{circulant.numeric}
\alias{circulant.numeric}
\title{circulant.numeric function}
\usage{
\method{circulant}{numeric}(x, ...)
}
\arguments{
  \item{x}{an numeric object}

  \item{...}{additional arguments}
}
\value{
a circulant matrix with base x
}
\description{
returns a circulant matrix with base x
}

