% Generated by roxygen2 (4.0.1): do not edit by hand
\name{circulant}
\alias{circulant}
\title{circulant function}
\usage{
circulant(x, ...)
}
\arguments{
\item{x}{an object}

\item{...}{additional arguments}
}
\value{
method circulant
}
\description{
generic function for constructing circulant matrices
}

