% Generated by roxygen2 (4.0.1): do not edit by hand
\name{cumbasehazard.basehazardspec}
\alias{cumbasehazard.basehazardspec}
\title{cumbasehazard.basehazardspec function}
\usage{
\method{cumbasehazard}{basehazardspec}(obj, ...)
}
\arguments{
\item{obj}{an object of class basehazardspec}

\item{...}{additional arguments -- currently there are none, but this is for extensibility}
}
\value{
a function returning the cumulative baseline hazard
}
\description{
A function to retrieve the cumulative baseline hazard function
}
\seealso{
\link{exponentialHaz}, \link{weibullHaz}, \link{gompertzHaz}, \link{makehamHaz}, \link{tpowHaz}
}

