% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splineFunctions.R
\name{cumulativeBspline.construct}
\alias{cumulativeBspline.construct}
\title{cumulativeBspline.construct function}
\usage{
cumulativeBspline.construct(x, basis)
}
\arguments{
\item{x}{a vector, the data}

\item{basis}{an object created by the getBbasis function}
}
\value{
an object that allows the integral of a given B-spline curve to be computed
}
\description{
A function to construct the integral of a B-spline curve given data and basis coefficients. Used in evaluating the cumulative baseline hazard.
}
