\name{density_exp}
\alias{density_exp}
\title{density_exp function}
\usage{
density_exp(inputs)
}
\arguments{
  \item{inputs}{inputs for the function including the model
  matrix, frailties, fixed effects and the parameters of
  the baseline hazard derived from this model}
}
\value{
the density function for the individual
}
\description{
A function to compute the density function for an
individual where the baseline hazard comes from an
exponential survival model
}

