\name{density_weibull}
\alias{density_weibull}
\title{density_weibull function}
\usage{
density_weibull(inputs)
}
\arguments{
  \item{inputs}{inputs for the function including the model
  matrix, frailties, fixed effects and the parameters of
  the baseline hazard derived from this model}
}
\value{
the density function for the individual
}
\description{
A function to compute the density function for an
individual where the baseline hazard comes from a Weibull
survival model
}

