\name{densityquantile_exp}
\alias{densityquantile_exp}
\title{densityquantile_exp function}
\usage{
densityquantile_exp(inputs)
}
\arguments{
  \item{inputs}{inputs for the function including the model
  matrix, frailties, fixed effects and the parameters of
  the baseline hazard derived from this model}
}
\value{
quantiles of the density function for the individual
}
\description{
A function to compute quantiles of the density function for
an individual where the baseline hazard comes from an
exponential survival model
}

