% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocessing.R
\name{densityquantile_PP}
\alias{densityquantile_PP}
\title{densityquantile_PP function}
\usage{
densityquantile_PP(inputs)
}
\arguments{
\item{inputs}{inputs for the function including the model matrix, frailties, fixed effects and the parameters of the baseline hazard derived from this model}
}
\value{
quantiles of the density function for the individual
}
\description{
A function to compute quantiles of the density function
}
