\name{densityquantile_weibull}
\alias{densityquantile_weibull}
\title{densityquantile_weibull function}
\usage{
densityquantile_weibull(inputs)
}
\arguments{
  \item{inputs}{inputs for the function including the model
  matrix, frailties, fixed effects and the parameters of
  the baseline hazard derived from this model}
}
\value{
quantiles of the density function for the individual
}
\description{
A function to compute quantiles of the density function for
an individual where the baseline hazard comes from a
Weibull survival model
}

