% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseHazST_prior.R
\name{derivindepGaussianpriorST}
\alias{derivindepGaussianpriorST}
\title{derivindepGaussianpriorST function}
\usage{
derivindepGaussianpriorST(beta = NULL, omega = NULL, eta = NULL, priors)
}
\arguments{
\item{beta}{X}

\item{omega}{X}

\item{eta}{X}

\item{priors}{X}
}
\value{
...
}
\description{
A function to
}
