\name{derivlogindepGaussianprior}
\alias{derivlogindepGaussianprior}
\title{derivlogindepGaussianprior function}
\usage{
derivlogindepGaussianprior(beta = NULL, omega = NULL, eta = NULL, priors)
}
\arguments{
  \item{beta}{X}

  \item{omega}{X}

  \item{eta}{X}

  \item{priors}{X}
}
\value{
...
}
\description{
A function to
}

