% Generated by roxygen2 (4.0.1): do not edit by hand
\name{distinfo.weibull}
\alias{distinfo.weibull}
\title{distinfo.weibull function}
\usage{
distinfo.weibull()
}
\value{
the transformation and inverse transformation, jacobian and hessian
}
\description{
A function to return the internal transformation function (and its inverse) for the baseline hazard type. E.g. for an Exponential baseline hazard, we work with the log rate, so log is the transformation function.
}

