% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.onAttach}
\alias{.onAttach}
\title{.onAttach function}
\usage{
.onAttach(libname, pkgname)
}
\arguments{
\item{libname}{libname argument}

\item{pkgname}{pkgname argument}
}
\value{
...
}
\description{
A function to print a welcome message on loading package
}
