% Generated by roxygen2 (4.0.1): do not edit by hand
\name{estimateY}
\alias{estimateY}
\title{estimateY function}
\usage{
estimateY(X, betahat, omegahat, surv, control)
}
\arguments{
\item{X}{X}

\item{betahat}{X}

\item{omegahat}{X}

\item{surv}{X}

\item{control}{X}
}
\value{
...
}
\description{
A function to
}

