% Generated by roxygen2 (4.0.1): do not edit by hand
\name{etapriorGauss}
\alias{etapriorGauss}
\title{etapriorGauss function}
\usage{
etapriorGauss(mean, sd)
}
\arguments{
\item{mean}{the prior mean, a vector of length 1 or more. 1 implies a common mean.}

\item{sd}{the prior standard deviation, a vector of length 1 or more. 1 implies a common standard deviation.}
}
\value{
an object of class "etapriorGauss"
}
\description{
A function to define Gaussian priors for eta.
}

