\name{EvalCov}
\alias{EvalCov}
\title{EvalCov function}
\usage{
EvalCov(cov.model, u, parameters)
}
\arguments{
  \item{cov.model}{an object of class covmodel}

  \item{u}{vector of distances}

  \item{parameters}{vector of parameters}
}
\value{
method EvalCov
}
\description{
Evaluate the covariance function
}

