% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ExponentialCovFct}
\alias{ExponentialCovFct}
\title{ExponentialCovFct function}
\usage{
ExponentialCovFct()
}
\value{
the exponential covariance function
}
\description{
A function to declare and also evaluate an exponential covariance function.
}

