% Generated by roxygen2 (4.0.1): do not edit by hand
\name{fixedpars}
\alias{fixedpars}
\title{fixedpars function}
\usage{
fixedpars(x)
}
\arguments{
\item{x}{an object of class mcmcspatsurv}
}
\value{
the mcmc chains
}
\description{
A function to return the mcmc chains for the covariate effects
}

