\name{frailtylag1}
\alias{frailtylag1}
\title{frailtylag1 function}
\usage{
frailtylag1(object, ...)
}
\arguments{
  \item{object}{an object of class mcmcspatsurv}

  \item{...}{other arguments to be passed to the plot
  function}
}
\value{
the lag 1 autocorrelation for each of the spatially
correlated frailty chains
}
\description{
A function to produce and return the lag 1 autocorrelation
for each of the spatially correlated frailty chains
}

