% Generated by roxygen2 (4.0.1): do not edit by hand
\name{GammafromY}
\alias{GammafromY}
\title{GammafromY function}
\usage{
GammafromY(Y, rootQeigs, mu)
}
\arguments{
\item{Y}{Y matrix}

\item{rootQeigs}{square root of the eigenvectors of the precision matrix}

\item{mu}{parameter of the latent Gaussian field}
}
\value{
Gamma
}
\description{
A function to change Ys (spatially correlated noise) into Gammas (white noise). Used in the MALA algorithm.
}

