% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GMRF.R
\name{GammaFromY_SPDE}
\alias{GammaFromY_SPDE}
\title{GammaFromY_SPDE function}
\usage{
GammaFromY_SPDE(Y, U, mu)
}
\arguments{
\item{Y}{Y}

\item{U}{upper Cholesky matrix}

\item{mu}{the mean}
}
\value{
the value of Gamma for the given Y
}
\description{
A function to go from Y to Gamma
}
\references{
\enumerate{
    \item Benjamin M. Taylor. Auxiliary Variable Markov Chain Monte Carlo for Spatial Survival and Geostatistical Models. Benjamin M. Taylor. Submitted. \url{http://arxiv.org/abs/1501.01665}
    \item Finn Lindgren, Havard Rue, Johan Lindstrom. An explicit link between Gaussian fields and Gaussian Markov random fields: the stochastic partial differential equation approach. Journal of the Royal Statistical Society: Series B 73(4)
}
}

