% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plottingfunctions.R
\name{getBackground}
\alias{getBackground}
\title{getBackground function}
\usage{
getBackground(poly, type = "stamen-toner")
}
\arguments{
\item{poly}{X}

\item{type}{X}
}
\value{
...
}
\description{
A function to
}

