% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMRF.R
\name{getbb}
\alias{getbb}
\title{getbb function}
\usage{
getbb(obj)
}
\arguments{
\item{obj}{a spatial object e.g. a SpatialPolygonsDataFrame, SpatialPolygons, etc ... anything with a bounding box that can be computed with bbox(obj)}
}
\value{
a SpatialPolygons object: the bounding box
}
\description{
A function to get the bounding box of a Spatial object
}
