% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splineFunctions.R
\name{getBbasis}
\alias{getBbasis}
\title{getBbasis function}
\usage{
getBbasis(x, knots, degree, force = FALSE)
}
\arguments{
\item{x}{a vector of data}

\item{knots}{a vector of knots in ascending order. The first and last knots must be respectively the minimum and maximum of x.}

\item{degree}{the degree of the spline}

\item{force}{logical: skip check on knots? (not recommended!)}
}
\value{
the knots and the piecewise polynomial coefficients for a B-spline basis function i.e. the basis functions.
}
\description{
A function returning the piecewise polynomial coefficients for a B-spline basis function i.e. the basis functions.
}
