% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMRF.R
\name{getgrd}
\alias{getgrd}
\title{getgrd function}
\usage{
getgrd(shape, cellwidth)
}
\arguments{
\item{shape}{an object of class SpatialPolygons or SpatialPolygonsDataFrame}

\item{cellwidth}{a scalar, the width of the grid cells}
}
\value{
a SpatialPolygons object: the grid on which prediction of the spatial effects will occur
}
\description{
A function to create a regular grid over an observation window in order to model the spatial randome effects as a Gaussian
Markov random field.
}
\references{
\enumerate{
    \item Benjamin M. Taylor. Auxiliary Variable Markov Chain Monte Carlo for Spatial Survival and Geostatistical Models. Benjamin M. Taylor. Submitted. \url{http://arxiv.org/abs/1501.01665}
    \item Finn Lindgren, Havard Rue, Johan Lindstrom. An explicit link between Gaussian fields and Gaussian Markov random fields: the stochastic partial differential equation approach. Journal of the Royal Statistical Society: Series B 73(4)
}
}

