% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocessing.R
\name{getGrid}
\alias{getGrid}
\title{getGrid function}
\usage{
getGrid(mod, returnclass = "SpatialPolygonsDataFrame")
}
\arguments{
\item{mod}{an object of class mcmcspatsurv, returned by the function survspat}

\item{returnclass}{the class of object to return, default is a'SpatialPolygonsDataFrame'. Other options are 'raster', which returns a raster brick; or 'SpatialPixelsDataFrame'}
}
\value{
a SpatialPolygonsDataFrame in which Monte Carlo expectations can be stored and later plotted.
}
\description{
A function to extract and return the computational grid from a gridded analysis.
}
