% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatsurvMisc.R
\name{getleneta}
\alias{getleneta}
\title{getleneta function}
\usage{
getleneta(cov.model)
}
\arguments{
\item{cov.model}{a covariance model}
}
\value{
the length of eta
}
\description{
A function to compute the length of eta
}
