\name{getsurvdata}
\alias{getsurvdata}
\title{getsurvdata function}
\usage{
getsurvdata(x)
}
\arguments{
  \item{x}{an object of class mcmcspatsurv}
}
\value{
the survival data from an object of class mcmcspatsurv
}
\description{
A function to return the survival data from an object of
class mcmcspatsurv
}

