% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basehazardspec.R
\name{gradcumbasehazard.basehazardspec}
\alias{gradcumbasehazard.basehazardspec}
\title{gradcumbasehazard.basehazardspec function}
\usage{
\method{gradcumbasehazard}{basehazardspec}(obj, ...)
}
\arguments{
\item{obj}{an object of class basehazardspec}

\item{...}{additional arguments -- currently there are none, but this is for extensibility}
}
\value{
a function returning the gradient of the cumulative baseline hazard
}
\description{
A function to retrieve the gradient of the cumulative baseline hazard function
}
\seealso{
\link{exponentialHaz}, \link{weibullHaz}, \link{gompertzHaz}, \link{makehamHaz}, \link{tpowHaz}
}
