% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatgrid.R
\name{gridY}
\alias{gridY}
\title{gridY function}
\usage{
gridY(Y, control)
}
\arguments{
\item{Y}{estimate of Y}

\item{control}{control parameters}
}
\value{
...
}
\description{
A function to put estimated individual Y's onto a grid
}
