\name{hazard_exp}
\alias{hazard_exp}
\title{hazard_exp function}
\usage{
hazard_exp(inputs)
}
\arguments{
  \item{inputs}{inputs for the function including the model
  matrix, frailties, fixed effects and the parameters of
  the baseline hazard derived from this model}
}
\value{
the hazard function for the individual
}
\description{
A function to compute the hazard function for an individual
where the baseline hazard comes from an exponential
survival model
}

