% Generated by roxygen2 (4.0.1): do not edit by hand
\name{hazard_PP}
\alias{hazard_PP}
\title{hazard_PP function}
\usage{
hazard_PP(inputs)
}
\arguments{
\item{inputs}{inputs for the function including the model matrix, frailties, fixed effects and the parameters of the baseline hazard derived from this model}
}
\value{
the hazard function for the individual
}
\description{
A function to compute the hazard function
}

