% Generated by roxygen2 (4.0.1): do not edit by hand
\name{hazardpars}
\alias{hazardpars}
\title{hazardpars function}
\usage{
hazardpars(x)
}
\arguments{
\item{x}{an object of class mcmcspatsurv}
}
\value{
the mcmc chains
}
\description{
A function to return the mcmc chains for the hazard function parameters
}

