% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_imputation.R
\name{imputationModel}
\alias{imputationModel}
\title{imputationModel function}
\usage{
imputationModel(formula, offset, covariateData, priors)
}
\arguments{
\item{formula}{X}

\item{offset}{X}

\item{covariateData}{X}

\item{priors}{X}
}
\value{
...
}
\description{
A function to
}
