% Generated by roxygen2 (4.0.1): do not edit by hand
\name{inference.control}
\alias{inference.control}
\title{inference.control function}
\usage{
inference.control(gridded = FALSE, cellwidth = NULL, ext = 2,
  optimcontrol = NULL, hessian = FALSE, plotcal = FALSE,
  timeonlyMCMC = FALSE)
}
\arguments{
\item{gridded}{logical. Whether to perform compuation on a grid. Default is FALSE.}

\item{cellwidth}{the width of computational cells to use}

\item{ext}{integer the number of times to extend the computational grid by in order to perform compuitation. The default is 2.}

\item{optimcontrol}{a list of optional arguments to be passed to optim for non-spatial models}

\item{hessian}{whether to return a numerical hessian. Set this to TRUE for non-spatial models.
equal to the number of parameters of the baseline hazard}

\item{plotcal}{logical, whether to produce plots of the MCMC calibration process, this is a technical option and should onyl be set
to TRUE if poor mixing is evident (the printed h is low), then it is also useful to use a graphics device with multiple plotting windows.}

\item{timeonlyMCMC}{logical, whether to only time the MCMC part of the algorithm, or whether to include in the reported running time the time taken to calibrate the method (default)}
}
\value{
returns parameters to be used in the function survspat
}
\description{
A function to control inferential settings. This function is used to set parameters for more advanced use of spatsurv.
}
\seealso{
\link{survspat}
}

