\name{iteration}
\alias{iteration}
\title{iteration number}
\usage{
iteration(obj)
}
\arguments{
  \item{obj}{an mcmc iterator}
}
\value{
integer iteration number, starting from 1.
}
\description{
within a loop, this is the iteration number we are
currently doing.
}
\details{
get the iteration number
}

