% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mcmcProgressNone}
\alias{mcmcProgressNone}
\title{null progress monitor}
\usage{
mcmcProgressNone(mcmcloop)
}
\arguments{
\item{mcmcloop}{an mcmc loop iterator}
}
\value{
a progress monitor
}
\description{
a progress monitor that does nothing
}

