% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mcmcProgressPrint}
\alias{mcmcProgressPrint}
\title{printing progress monitor}
\usage{
mcmcProgressPrint(mcmcloop)
}
\arguments{
\item{mcmcloop}{an mcmc loop iterator}
}
\value{
a progress monitor
}
\description{
a progress monitor that prints each iteration
}

