% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcProgress.R
\name{mcmcProgressTextBar}
\alias{mcmcProgressTextBar}
\title{text bar progress monitor}
\usage{
mcmcProgressTextBar(mcmcloop)
}
\arguments{
\item{mcmcloop}{an mcmc loop iterator}
}
\value{
a progress monitor
}
\description{
a progress monitor that uses a text progress bar
}
