% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splineFunctions.R
\name{midpts}
\alias{midpts}
\title{midpts function}
\usage{
midpts(x)
}
\arguments{
\item{x}{a vector}
}
\value{
the midpoints, a vector of length length(x)-1
}
\description{
A function to compute the midpoints of a vector
}
