% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiWayHaz.R
\name{multiWayHaz}
\alias{multiWayHaz}
\title{multiWayHaz function}
\usage{
multiWayHaz(bhlist, bhtime, bhfix, MLinits = NULL)
}
\arguments{
\item{bhlist}{X}

\item{bhtime}{X}

\item{bhfix}{X}

\item{MLinits}{X}
}
\value{
...
}
\description{
A function to
}
