% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMRF.R
\name{neighLocs}
\alias{neighLocs}
\title{neighLocs function}
\usage{
neighLocs(coord, cellwidth, order)
}
\arguments{
\item{coord}{coordinate of interest}

\item{cellwidth}{a scalar, the width of the grid cells}

\item{order}{the order of the SPDE approximation: see Lindgren et al 2011 for details}
}
\value{
coordinates of centroids of neighbours
}
\description{
A function used in the computation of neighbours on non-rectangular grids. Not intended for general use.
}
\references{
\enumerate{
    \item Benjamin M. Taylor. Auxiliary Variable Markov Chain Monte Carlo for Spatial Survival and Geostatistical Models. Benjamin M. Taylor. Submitted. \url{http://arxiv.org/abs/1501.01665}
    \item Finn Lindgren, Havard Rue, Johan Lindstrom. An explicit link between Gaussian fields and Gaussian Markov random fields: the stochastic partial differential equation approach. Journal of the Royal Statistical Society: Series B 73(4)
}
}

