% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GMRF.R
\name{neighOrder}
\alias{neighOrder}
\title{neighOrder function}
\usage{
neighOrder(neighlocs)
}
\arguments{
\item{neighlocs}{an object created by the function neighLocs}
}
\value{
the neighbour orders
}
\description{
A function to compute the order of a set of neighbours. Not intended for general use.
}
\references{
\enumerate{
    \item Benjamin M. Taylor. Auxiliary Variable Markov Chain Monte Carlo for Spatial Survival and Geostatistical Models. Benjamin M. Taylor. Submitted. \url{http://arxiv.org/abs/1501.01665}
    \item Finn Lindgren, Havard Rue, Johan Lindstrom. An explicit link between Gaussian fields and Gaussian Markov random fields: the stochastic partial differential equation approach. Journal of the Royal Statistical Society: Series B 73(4)
}
}

