% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseHazST_prior.R
\name{omegapriorGaussST}
\alias{omegapriorGaussST}
\title{omegapriorGaussST function}
\usage{
omegapriorGaussST(basehaz, fmean, fsd, taumean, tausd, thetamean, thetasd)
}
\arguments{
\item{basehaz}{X}

\item{fmean}{X}

\item{fsd}{X}

\item{taumean}{X}

\item{tausd}{X}

\item{thetamean}{X}

\item{thetasd}{X}
}
\value{
...
}
\description{
A function to
}
