% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optifix.R
\name{optifix}
\alias{optifix}
\title{optifix function}
\usage{
optifix(
  par,
  fixed,
  fn,
  gr = NULL,
  ...,
  method = c("Nelder-Mead", "BFGS", "CG", "L-BFGS-B", "SANN"),
  lower = -Inf,
  upper = Inf,
  control = list(),
  hessian = FALSE
)
}
\arguments{
\item{par}{X}

\item{fixed}{X}

\item{fn}{X}

\item{gr}{X}

\item{...}{X}

\item{method}{X}

\item{lower}{X}

\item{upper}{X}

\item{control}{X}

\item{hessian}{X}
}
\value{
...
}
\description{
optifix. Optimise with fixed parameters
}
\details{
its like optim, but with fixed parameters.

specify a second argument 'fixed', a vector of TRUE/FALSE values.
If TRUE, the corresponding parameter in fn() is fixed. Otherwise its
variable and optimised over.

The return thing is the return thing from optim() but with a couple of extra
bits - a vector of all the parameters and a vector copy of the 'fixed' argument.

Written by Barry Rowlingson <b.rowlingson@lancaster.ac.uk> October 2011

This file released under a CC By-SA license:
http://creativecommons.org/licenses/by-sa/3.0/

and must retain the text: "Originally written by Barry Rowlingson" in comments.
}
