% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatgrid.R
\name{plot.FFTgrid}
\alias{plot.FFTgrid}
\title{plot.FFTgrid function}
\usage{
\method{plot}{FFTgrid}(x, y = NULL, ...)
}
\arguments{
\item{x}{X}

\item{y}{X}

\item{...}{X}
}
\value{
...
}
\description{
A function to
}
