% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splineFunctions.R
\name{polyadd}
\alias{polyadd}
\title{polyadd function}
\usage{
polyadd(poly1, poly2)
}
\arguments{
\item{poly1}{a vector of coefficients for the first polynomial of length degree plus 1}

\item{poly2}{a vector of coefficients for the second polynomial of length degree plus 1}
}
\value{
the coefficients of the sum of poly1 and poly2
}
\description{
A function to add two polynomials in the form of vectors of coefficients. The first
element of the vector being the constant (order 0) term
}
