\name{print.mcmcspatsurv}
\alias{print.mcmcspatsurv}
\title{print.mcmcspatsurv function}
\usage{
\method{print}{mcmcspatsurv}(x, probs = c(0.5, 0.025, 0.975), digits = 3,
  scientific = -3, ...)
}
\arguments{
  \item{x}{an object inheriting class mcmcspatsurv}

  \item{probs}{vector of quantiles to return}

  \item{digits}{see help file for format}

  \item{scientific}{see help file for format}

  \item{...}{additional arguments}
}
\value{
summary tables
}
\description{
A function to print summary tables from an MCMC run
}

