% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocessing.R
\name{print.mlspatsurv}
\alias{print.mlspatsurv}
\title{print.mlspatsurv function}
\usage{
\method{print}{mlspatsurv}(x, probs = c(0.5, 0.025, 0.975), digits = 3, scientific = -3, ...)
}
\arguments{
\item{x}{an object inheriting class mcmcspatsurv}

\item{probs}{vector of quantiles to return}

\item{digits}{see help file ?format}

\item{scientific}{see help file ?format}

\item{...}{additional arguments, not used here}
}
\value{
prints summary tables to the console
}
\description{
A function to print summary tables from an MCMC run
}
\seealso{
\link{quantile.mcmcspatsurv}, \link{summary.mcmcspatsurv}, \link{vcov.mcmcspatsurv},
\link{frailtylag1}, \link{spatialpars}, \link{hazardpars}, \link{fixedpars}, \link{randompars},
\link{baselinehazard}, \link{predict.mcmcspatsurv}, \link{priorposterior}, \link{posteriorcov}, \link{MCE},
\link{hazardexceedance}
}
