% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputSummary.R
\name{print.textSummary}
\alias{print.textSummary}
\title{print.textSummary function}
\usage{
\method{print}{textSummary}(x, ...)
}
\arguments{
\item{x}{an object inheriting class textSummary}

\item{...}{additional arguments, not used here}
}
\value{
prints a text summary of 'x' to the console
}
\description{
A function to print summary tables from an MCMC run
}
