% Generated by roxygen2 (4.0.1): do not edit by hand
\name{priorposterior}
\alias{priorposterior}
\title{priorposterior function}
\usage{
priorposterior(x, breaks = 30, ylab = "Density", main = "",
  pause = TRUE, ...)
}
\arguments{
\item{x}{X}

\item{breaks}{X}

\item{ylab}{X}

\item{main}{X}

\item{pause}{X}

\item{...}{X}
}
\value{
...
}
\description{
A function to
}

