\name{proposalvariance.exp}
\alias{proposalvariance.exp}
\title{proposalvariance.exp function}
\usage{
proposalvariance.exp(X, delta, tm, betahat, omegahat, Yhat, priors, covmodel, u)
}
\arguments{
  \item{X}{X}

  \item{delta}{X}

  \item{tm}{X}

  \item{betahat}{X}

  \item{omegahat}{X}

  \item{Yhat}{X}

  \item{priors}{X}

  \item{covmodel}{X}

  \item{u}{X}
}
\value{
...
}
\description{
A function to
}

