% Generated by roxygen2 (4.0.1): do not edit by hand
\name{proposalVariance}
\alias{proposalVariance}
\title{proposalVariance function}
\usage{
proposalVariance(X, surv, betahat, omegahat, Yhat, priors, cov.model, u,
  control)
}
\arguments{
\item{X}{X}

\item{surv}{X}

\item{betahat}{X}

\item{omegahat}{X}

\item{Yhat}{X}

\item{priors}{X}

\item{cov.model}{X}

\item{u}{X}

\item{control}{X}
}
\value{
...
}
\description{
A function to
}

