\name{proposalvariance.weibull}
\alias{proposalvariance.weibull}
\title{proposalvariance.weibull function}
\usage{
proposalvariance.weibull(X, delta, tm, betahat, omegahat, Yhat, priors,
  covmodel, u)
}
\arguments{
  \item{X}{X}

  \item{delta}{X}

  \item{tm}{X}

  \item{betahat}{X}

  \item{omegahat}{X}

  \item{Yhat}{X}

  \item{priors}{X}

  \item{covmodel}{X}

  \item{u}{X}
}
\value{
...
}
\description{
A function to
}

