% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibratemcmc.R
\name{proposalVariance_SPDE}
\alias{proposalVariance_SPDE}
\title{proposalVariance_SPDE function}
\usage{
proposalVariance_SPDE(
  X,
  surv,
  betahat,
  omegahat,
  Yhat,
  priors,
  cov.model,
  u,
  control
)
}
\arguments{
\item{X}{the design matrix, containing covariate information}

\item{surv}{an object of class Surv}

\item{betahat}{an estimate of beta}

\item{omegahat}{an estimate of omega}

\item{Yhat}{an estimate of Y}

\item{priors}{the priors}

\item{cov.model}{the spatial covariance model}

\item{u}{a vector of pairwise distances}

\item{control}{a list containg various control parameters for the MCMC and post-processing routines}
}
\value{
an estimate of eta and also an approximate scaling matrix for the MCMC
}
\description{
A function to compute an approximate scaling matrix for the MCMC algorithm. Not intended for general use.
}
