% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RWprior.R
\name{psplineRWprior}
\alias{psplineRWprior}
\title{psplineRWprior function}
\usage{
psplineRWprior(taumean, tausd, basehaz, order = 2)
}
\arguments{
\item{taumean}{the prior mean, a vector of length 1 or more. 1 implies a common mean.}

\item{tausd}{the prior standard deviation, a vector of length 1 or more. 1 implies a common standard deviation.}

\item{basehaz}{an object inheriting class "basehazardspec", specificlly, this function was used for such objects created by a call to the function PsplineHaz}

\item{order}{the order of the random walk, default is 2}
}
\value{
an object of class "omegapriorGauss"
}
\description{
A function to define Gaussian priors for omega. This function simply stores a vector of means and standard deviations to be passed to the main MCMC function, survspat.
}
\seealso{
\link{survspat}, \link{betapriorGauss}, \link{omegapriorGauss}, \link{etapriorGauss}, \link{indepGaussianprior}, \link{derivindepGaussianprior}
}
