\name{quantile.mcmcspatsurv}
\alias{quantile.mcmcspatsurv}
\title{quantile.mcmcspatsurv function}
\usage{
\method{quantile}{mcmcspatsurv}(x, probs = c(0.025, 0.5, 0.975), ...)
}
\arguments{
  \item{x}{an object of class mcmc spatsurv}

  \item{probs}{vector of probabilities}

  \item{...}{other arguments to be passed to the function}
}
\value{
...
}
\description{
A function to extract quantiles of the parameters from an
mcmc run
}

