% Generated by roxygen2 (4.0.1): do not edit by hand
\name{randompars}
\alias{randompars}
\title{randompars function}
\usage{
randompars(x)
}
\arguments{
\item{x}{an object of class mcmcspatsurv}
}
\value{
the mcmc chains
}
\description{
A function to return the mcmc chains for the spatially correlated frailties
}

