% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocessing.R
\name{reconstruct.bs}
\alias{reconstruct.bs}
\title{reconstruct.bs function}
\usage{
reconstruct.bs(mod, ...)
}
\arguments{
\item{mod}{an object}

\item{...}{additional arguments}
}
\value{
method reconstruct.bs
}
\description{
Generic function for reconstructing B-spline covariate effects. See ?reconstruct.bs.mcmcspatsurv and ?reconstruct.bs.coxph
}
