% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocessing.R
\name{residuals.mcmcspatsurv}
\alias{residuals.mcmcspatsurv}
\title{resuiduals.mcmcspatsurv function}
\usage{
\method{residuals}{mcmcspatsurv}(object, type = "Cox-Snell", ...)
}
\arguments{
\item{object}{an object produced by the function survspat}

\item{type}{type of residuals to return. Possible choices are 'Cox-Snell', 'modified-Cox-Snell', 'Martingale' or 'deviance'.}

\item{...}{other arguments (not used here)}
}
\value{
the residuals
}
\description{
A function to compute Cox-Snell / modeified Cox-Snell / Martingale or Deviance residuals
}
