% Generated by roxygen2 (4.0.1): do not edit by hand
\name{setupHazard}
\alias{setupHazard}
\title{setupHazard function}
\usage{
setupHazard(dist, pars, grad = FALSE, hess = FALSE)
}
\arguments{
\item{dist}{X}

\item{pars}{X}

\item{grad}{X}

\item{hess}{X}
}
\value{
...
}
\description{
A function to
}

