% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMRF.R
\name{setupPrecMatStruct}
\alias{setupPrecMatStruct}
\title{setupPrecMatStruct function}
\usage{
setupPrecMatStruct(shape, cellwidth, no)
}
\arguments{
\item{shape}{an object of class SpatialPolygons or SpatialPolygonsDataFrame}

\item{cellwidth}{a scalar, the width of the grid cells}

\item{no}{the order of the SPDE approximation: see Lindgren et al 2011 for details}
}
\value{
the computational grid and a function for constructing the precision matrix
}
\description{
A function to set up the computational grid and precision matrix structure for SPDE models.
}
\references{
\enumerate{
    \item Benjamin M. Taylor. Auxiliary Variable Markov Chain Monte Carlo for Spatial Survival and Geostatistical Models. Benjamin M. Taylor. Submitted. \url{http://arxiv.org/abs/1501.01665}
    \item Finn Lindgren, Havard Rue, Johan Lindstrom. An explicit link between Gaussian fields and Gaussian Markov random fields: the stochastic partial differential equation approach. Journal of the Royal Statistical Society: Series B 73(4)
}
}

