% Generated by roxygen2 (4.0.1): do not edit by hand
\name{spatialpars}
\alias{spatialpars}
\title{spatialpars function}
\usage{
spatialpars(x)
}
\arguments{
\item{x}{an object of class mcmcspatsurv}
}
\value{
the mcmc chains
}
\description{
A function to return the mcmc chains for the spatial covariance function parameters
}

