% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatsurv.R
\docType{package}
\name{spatsurv-package}
\alias{spatsurv}
\alias{spatsurv-package}
\title{spatsurv}
\format{An object of class \code{logical} of length 1.}
\usage{
spatsurv
}
\description{
An R package for spatially correlated parametric proportional hazards survial analysis.
}
\details{
\tabular{ll}{
Package: \tab spatsurv\cr
Version: \tab 0.9-11\cr
Date: \tab 2015-06-24\cr
License: \tab GPL-3 \cr
}



section{Dependencies}{
The package \code{spatsurv} depends upon some other important contributions to CRAN in order to operate; their uses here are indicated:\cr\cr
    survival, sp, spatstat, raster, iterators, RandomFields, fields, rgl, Matrix, stringr, RColorBrewer, geostatsp, rgeos.
}

section{Citation}{
To cite use of \code{spatsurv}, the user may refer to the following work:\cr\cr
Benjamin M. Taylor and Barry S. Rowlingson (2017).\cr
spatsurv: An R Package for Bayesian Inference with Spatial Survival Models.\cr
Journal of Statistical Software, 77(4), 1-32, doi:10.18637/jss.v077.i04.
}

references{
X
}
}
\author{
Benjamin Taylor, Health and Medicine, Lancaster University,
 Barry Rowlingson, Health and Medicine, Lancaster University
}
\keyword{package}

