% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatsurvMisc.R
\name{spatsurvVignette}
\alias{spatsurvVignette}
\title{spatsurvVignette function}
\usage{
spatsurvVignette()
}
\value{
displays the vignette by calling browseURL
}
\description{
Display the introductory vignette for the spatsurv package.
}
