% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covarianceFunctions.R
\name{SPDE}
\alias{SPDE}
\title{SPDE function}
\usage{
SPDE(ord)
}
\arguments{
\item{ord}{the order of the model to be used, currently an integer between 1 an 3. See Lindgren 2011 paper.}
}
\value{
an covariance function based on the SPDE model
}
\description{
A function to declare and evaluate an SPDE covariance function.
}
\seealso{
\link{ExponentialCovFct}, \link{covmodel}, CovarianceFct
}
