% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GMRF.R
\name{SPDEprec}
\alias{SPDEprec}
\title{SPDEprec function}
\usage{
SPDEprec(a, ord)
}
\arguments{
\item{a}{parameter a, see Lindgren et al 2011.}

\item{ord}{the order of the SPDE model, see Lindgren et al 2011.}
}
\value{
a function used for creating the precision matrix
}
\description{
A function to used in entering elements into the precision matrix of an SPDE model. Not intended for general use.
}
\references{
\enumerate{
    \item Benjamin M. Taylor. Auxiliary Variable Markov Chain Monte Carlo for Spatial Survival and Geostatistical Models. Benjamin M. Taylor. Submitted. \url{http://arxiv.org/abs/1501.01665}
    \item Finn Lindgren, Havard Rue, Johan Lindstrom. An explicit link between Gaussian fields and Gaussian Markov random fields: the stochastic partial differential equation approach. Journal of the Royal Statistical Society: Series B 73(4)
}
}

