\name{SpikedExponentialCovFct}
\alias{SpikedExponentialCovFct}
\title{SpikedExponentialCovFct function}
\usage{
SpikedExponentialCovFct()
}
\value{
the spiked exponential covariance function
}
\description{
A function to declare and also evaluate a spiked
exponential covariance function. This is an exponential
covariance function with a nugget.
}
\seealso{
\link{ExponentialCovFct}, \link{covmodel},
\link{CovarianceFct}
}

