% Generated by roxygen2 (4.0.1): do not edit by hand
\name{summary.mcmcspatsurv}
\alias{summary.mcmcspatsurv}
\title{summary.mcmcspatsurv function}
\usage{
\method{summary}{mcmcspatsurv}(object, probs = c(0.5, 0.025, 0.975), ...)
}
\arguments{
\item{object}{an object inheriting class mcmcspatsurv}

\item{probs}{vector of quantiles to return}

\item{...}{additional arguments}
}
\value{
summary tables
}
\description{
A function to print summary tables from an MCMC run
}

