\name{survival_exp}
\alias{survival_exp}
\title{survival_exp function}
\usage{
survival_exp(inputs)
}
\arguments{
  \item{inputs}{inputs for the function including the model
  matrix, frailties, fixed effects and the parameters of
  the baseline hazard derived from this model}
}
\value{
the survival function for the individual
}
\description{
A function to compute the survival function for an
individual where the baseline hazard comes from an
exponential survival model
}

