% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocessing.R
\name{survival_PP}
\alias{survival_PP}
\title{survival_PP function}
\usage{
survival_PP(inputs)
}
\arguments{
\item{inputs}{inputs for the function including the model matrix, frailties, fixed effects and the parameters of the baseline hazard derived from this model}
}
\value{
the survival function for the individual
}
\description{
A function to compute an individual's survival function
}
