\name{survival_weibull}
\alias{survival_weibull}
\title{survival_weibull function}
\usage{
survival_weibull(inputs)
}
\arguments{
  \item{inputs}{inputs for the function including the model
  matrix, frailties, fixed effects and the parameters of
  the baseline hazard derived from this model}
}
\value{
the survival function for the individual
}
\description{
A function to compute the survival function for an
individual where the baseline hazard comes from a Weibull
survival model
}

