% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partial_likelihood.R
\name{timevaryingPL}
\alias{timevaryingPL}
\title{timevaryingPL function}
\usage{
timevaryingPL(formula, t0, dist, data, optimcontrol = NULL)
}
\arguments{
\item{formula}{X}

\item{t0}{X}

\item{dist}{X}

\item{data}{X}

\item{optimcontrol}{X}
}
\value{
...
}
\description{
A function to
}

