% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiWayHazAdditive.R
\name{TwoWayHazAdditive}
\alias{TwoWayHazAdditive}
\title{TwoWayHazAdditive function}
\usage{
TwoWayHazAdditive(bhlist, bhtime, bhfix, MLinits = NULL)
}
\arguments{
\item{bhlist}{X}

\item{bhtime}{X}

\item{bhfix}{X}

\item{MLinits}{X}
}
\value{
...
}
\description{
A function to
}
