% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maplayers.R
\name{urlTemplate}
\alias{urlTemplate}
\title{urlTemplate function}
\usage{
urlTemplate(name = "Stamen_Toner")
}
\arguments{
\item{name}{name of the template to use, the default is 'Stamen_Toner'}
}
\value{
url for the leaflet template
}
\description{
A function to return a url for a leaflet template for use as map backgrounds with the spplot1 function.
}
\details{
Possible templates are: OpenStreetMap_Mapnik, OpenStreetMap_BlackAndWhite, OpenStreetMap_DE, 
OpenStreetMap_France, OpenStreetMap_HOT, OpenTopoMap, Thunderforest_OpenCycleMap, Thunderforest_Transport, 
Thunderforest_Landscape, Thunderforest_Outdoors, OpenMapSurfer_Roads, OpenMapSurfer_Grayscale, Hydda_Full, 
Hydda_Base, MapQuestOpen_OSM, Stamen_Toner, Stamen_TonerBackground, Stamen_TonerLite, Stamen_Watercolor, 
Stamen_Terrain, Stamen_TerrainBackground, Stamen_TopOSMRelief, Esri_WorldStreetMap, Esri_WorldTopoMap, 
Esri_WorldImagery, Esri_WorldTerrain, Esri_WorldShadedRelief, Esri_WorldPhysical, Esri_OceanBasemap, 
Esri_NatGeoWorldMap, Esri_WorldGrayCanvas, Acetate_all, Acetate_terrain, HERE_satelliteDay, 
HERE_hybridDayMobile, HERE_hybridDay

See \url{http://leaflet-extras.github.io/leaflet-providers/preview/} for other leaflet templates
}
