% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMRF.R
\name{YFromGamma_SPDE}
\alias{YFromGamma_SPDE}
\title{YFromGamma_SPDE function}
\usage{
YFromGamma_SPDE(gamma, U, mu)
}
\arguments{
\item{gamma}{Gamma}

\item{U}{upper Cholesky matrix}

\item{mu}{the mean}
}
\value{
the value of Y for the given Gamma
}
\description{
A function to go from Gamma to Y
}
\references{
\enumerate{
    \item Benjamin M. Taylor. Auxiliary Variable Markov Chain Monte Carlo for Spatial Survival and Geostatistical Models. Benjamin M. Taylor. Submitted. \url{http://arxiv.org/abs/1501.01665}
    \item Finn Lindgren, Havard Rue, Johan Lindstrom. An explicit link between Gaussian fields and Gaussian Markov random fields: the stochastic partial differential equation approach. Journal of the Royal Statistical Society: Series B 73(4)
}
}

