% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basehazardspec.R
\name{basehazard.basehazardspec}
\alias{basehazard.basehazardspec}
\title{basehazard.basehazardspec function}
\usage{
\method{basehazard}{basehazardspec}(obj, ...)
}
\arguments{
\item{obj}{an object of class basehazardspec}

\item{...}{additional arguments -- currently there are none, but this is for extensibility}
}
\value{
a function returning the baseline hazard
}
\description{
A function to retrieve the baseline hazard function
}
\seealso{
\link{exponentialHaz}, \link{weibullHaz}, \link{gompertzHaz}, \link{makehamHaz}, \link{tpowHaz}
}
