% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basehazardspec.R
\name{basehazard}
\alias{basehazard}
\title{basehazard function}
\usage{
basehazard(obj, ...)
}
\arguments{
\item{obj}{an object}

\item{...}{additional arguments -- currently there are none, but this is for extensibility}
}
\value{
method basehazard
}
\description{
Generic function for computing the baseline hazard
}
\seealso{
\link{basehazard.basehazardspec}, \link{exponentialHaz}, \link{weibullHaz}, \link{gompertzHaz}, \link{makehamHaz}, \link{tpowHaz}
}
