% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcPriors.R
\name{betapriorGauss}
\alias{betapriorGauss}
\title{betapriorGauss function}
\usage{
betapriorGauss(mean, sd)
}
\arguments{
\item{mean}{the prior mean, a vector of length 1 or more. 1 implies a common mean.}

\item{sd}{the prior standard deviation, a vector of length 1 or more. 1 implies a common standard deviation.}
}
\value{
an object of class "betapriorGauss"
}
\description{
A function to define Gaussian priors for beta. This function simply stores a vector of means and standard deviations to be passed to the main MCMC function, survspat.
}
\seealso{
\link{survspat}, \link{betapriorGauss}, \link{omegapriorGauss}, \link{etapriorGauss}, \link{indepGaussianprior}, \link{derivindepGaussianprior}
}
