% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circulant.R
\name{circulant}
\alias{circulant}
\title{circulant function}
\usage{
circulant(x, ...)
}
\arguments{
\item{x}{an object}

\item{...}{additional arguments}
}
\value{
method circulant
}
\description{
generic function for constructing circulant matrices
}
