% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covarianceFunctions.R
\name{covmodel}
\alias{covmodel}
\title{covmodel function}
\usage{
covmodel(model, pars)
}
\arguments{
\item{model}{correlation type, a string see ?CovarianceFct}

\item{pars}{vector of additional parameters for certain classes of covariance function (eg Matern), these must be supplied in the order given in ?CovarianceFct and are not estimated}
}
\value{
an object of class covmodel
}
\description{
A function to define the spatial covariance model, see also ?CovarianceFct. Note that the parameters defined by the 'pars' argument are fixed,
i.e. not estimated by the MCMC algorithm. To have spatsurv estimate these parameters, the user must construct a new covariance function to do so, 
see the spatsurv vignette.
}
\seealso{
CovarianceFct
}
