\name{estimateY.weibull}
\alias{estimateY.weibull}
\title{estimateY.weibull function}
\usage{
estimateY.weibull(X, betahat, omegahat, tm, delta, u, covmodel)
}
\arguments{
  \item{X}{X}

  \item{betahat}{X}

  \item{omegahat}{X}

  \item{tm}{X}

  \item{delta}{X}

  \item{u}{X}

  \item{covmodel}{X}
}
\value{
...
}
\description{
A function to
}

