% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covarianceFunctions.R
\name{EvalCov}
\alias{EvalCov}
\title{EvalCov function}
\usage{
EvalCov(cov.model, u, parameters)
}
\arguments{
\item{cov.model}{an object of class covmodel}

\item{u}{vector of distances}

\item{parameters}{vector of parameters}
}
\value{
method EvalCov
}
\description{
This function is used to evaluate the covariance function within the MCMC run. Not intended for general use.
}
