% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covarianceFunctions.R
\name{ExponentialCovFct}
\alias{ExponentialCovFct}
\title{ExponentialCovFct function}
\usage{
ExponentialCovFct()
}
\value{
the exponential covariance function
}
\description{
A function to declare and also evaluate an exponential covariance function.
}
\seealso{
\link{SpikedExponentialCovFct}, \link{covmodel}, CovarianceFct
}
