% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocessing.R
\name{fixedpars}
\alias{fixedpars}
\title{fixedpars function}
\usage{
fixedpars(x)
}
\arguments{
\item{x}{an object of class mcmcspatsurv}
}
\value{
the beta mcmc chains
}
\description{
A function to return the mcmc chains for the covariate effects
}
\seealso{
\link{print.mcmcspatsurv}, \link{quantile.mcmcspatsurv}, \link{summary.mcmcspatsurv}, \link{vcov.mcmcspatsurv},
\link{frailtylag1}, \link{spatialpars}, \link{hazardpars}, \link{randompars},
\link{baselinehazard}, \link{predict.mcmcspatsurv}, \link{priorposterior}, \link{posteriorcov}, \link{MCE},
\link{hazardexceedance}
}
