% Generated by roxygen2 (4.0.1): do not edit by hand
\name{fixmatrix}
\alias{fixmatrix}
\title{fixmatrix function}
\usage{
fixmatrix(mat)
}
\arguments{
\item{mat}{a matrix}
}
\value{
the fixed matrix
}
\description{
!! THIS FUNCTION IS NOT INTENDED FOR GENERAL USE !!
}
\details{
A function to fix up an estimated covariance matrix using a VERY ad-hoc method.
}

