% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibratemcmc.R
\name{fixmatrix}
\alias{fixmatrix}
\title{fixmatrix function}
\usage{
fixmatrix(mat)
}
\arguments{
\item{mat}{a matrix}
}
\value{
the fixed matrix
}
\description{
!! THIS FUNCTION IS NOT INTENDED FOR GENERAL USE !!
}
\details{
A function to fix up an estimated covariance matrix using a VERY ad-hoc method.
}
