% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocessing.R
\name{frailtylag1}
\alias{frailtylag1}
\title{frailtylag1 function}
\usage{
frailtylag1(object, plot = TRUE, lag = 1, ...)
}
\arguments{
\item{object}{an object inheriting class mcmcspatsurv}

\item{plot}{logical whether to plot the result, default is TRUE}

\item{lag}{the lag to plot, the default is 1}

\item{...}{other arguments to be passed to the plot function}
}
\value{
the lag 1 autocorrelation for each of the spatially correlated frailty chains
}
\description{
A function to produce a plot of, and return, the lag 1 (or higher, see argument 'lag') autocorrelation for each of the spatially correlated frailty chains
}
\seealso{
\link{print.mcmcspatsurv}, \link{quantile.mcmcspatsurv}, \link{summary.mcmcspatsurv}, \link{vcov.mcmcspatsurv},
\link{spatialpars}, \link{hazardpars}, \link{fixedpars}, \link{randompars},
\link{baselinehazard}, \link{predict.mcmcspatsurv}, \link{priorposterior}, \link{posteriorcov}, \link{MCE},
\link{hazardexceedance}
}
