\name{fs}
\docType{data}
\alias{fs}
\title{London Fire Brigade property}
\description{London Fire Brigade property}
\usage{data(fs)}
\format{data.frame}
\source{http://data.london.gov.uk/}
\references{\url{http://data.london.gov.uk/},\url{http://www.nationalarchives.gov.uk/doc/open-government-licence/}}
\keyword{datasets}
\examples{fire <- data(fs)}
