% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingfunctions.R
\name{getBackground}
\alias{getBackground}
\title{getBackground function}
\usage{
getBackground(poly, type = "stamen-toner")
}
\arguments{
\item{poly}{a spatial object that can be transformed and the extent obtained using the bbox function.}

\item{type}{see ?openmap}
}
\value{
...
}
\description{
A function to
}
