% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMRF.R
\name{getgrd}
\alias{getgrd}
\title{getgrd function}
\usage{
getgrd(shape, cellwidth)
}
\arguments{
\item{shape}{an object of class SpatialPolygons or SpatialPolygonsDataFrame}

\item{cellwidth}{a scalar, the width of the grid cells}
}
\value{
a SpatialPolygons object: the grid on which prediction of the spatial effects will occur
}
\description{
A function to create a regular grid over an observation window in order to model the spatial randome effects as a Gaussian
Markov random field.
}
\references{
\enumerate{
    \item Benjamin M. Taylor and Barry S. Rowlingson (2017). spatsurv: An R Package for Bayesian Inference with Spatial Survival Models. Journal of Statistical Software, 77(4), 1-32, doi:10.18637/jss.v077.i04.
    \item Finn Lindgren, Havard Rue, Johan Lindstrom. An explicit link between Gaussian fields and Gaussian Markov random fields: the stochastic partial differential equation approach. Journal of the Royal Statistical Society: Series B 73(4)
}
}
