% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatsurvMisc.R
\name{getsurvdata}
\alias{getsurvdata}
\title{getsurvdata function}
\usage{
getsurvdata(x)
}
\arguments{
\item{x}{an object of class mcmcspatsurv}
}
\value{
the survival data from an object of class mcmcspatsurv
}
\description{
A function to return the survival data from an object of class mcmcspatsurv. This function is not intended for general use.
}
