% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocessing.R
\name{hazard_PP}
\alias{hazard_PP}
\title{hazard_PP function}
\usage{
hazard_PP(inputs)
}
\arguments{
\item{inputs}{inputs for the function including the model matrix, frailties, fixed effects and the parameters of the baseline hazard derived from this model}
}
\value{
the hazard function for the individual
}
\description{
A function to compute an individual's hazard function.
}
