% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocessing.R
\name{hazardpars}
\alias{hazardpars}
\title{hazardpars function}
\usage{
hazardpars(x)
}
\arguments{
\item{x}{an object of class mcmcspatsurv}
}
\value{
the omega mcmc chains
}
\description{
A function to return the mcmc chains for the hazard function parameters
}
\seealso{
\link{print.mcmcspatsurv}, \link{quantile.mcmcspatsurv}, \link{summary.mcmcspatsurv}, \link{vcov.mcmcspatsurv},
\link{frailtylag1}, \link{spatialpars}, \link{fixedpars}, \link{randompars},
\link{baselinehazard}, \link{predict.mcmcspatsurv}, \link{priorposterior}, \link{posteriorcov}, \link{MCE},
\link{hazardexceedance}
}
