\name{labelomegamatrix}
\alias{labelomegamatrix}
\title{labelomegamatrix function}
\usage{
labelomegamatrix(m, dist)
}
\arguments{
  \item{m}{a matrix}

  \item{dist}{distribution function of the baseline hazard}
}
\value{
a lebelled matrix
}
\description{
A function to label output matrices for the omegavariable
}

