% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcIterator.R
\name{mcmcpars}
\alias{mcmcpars}
\title{mcmcpars function}
\usage{
mcmcpars(nits, burn, thin, inits = NULL, adaptivescheme = NULL)
}
\arguments{
\item{nits}{numer of iterations,}

\item{burn}{length of burnin}

\item{thin}{thinning parameter eg operated on chain every 'thin' iteration (eg store output or compute some posterior functional)}

\item{inits}{NOT CURRENTLY IN USE}

\item{adaptivescheme}{NOT CURRENTLY IN USE}
}
\value{
mcmc parameters
}
\description{
A function for setting MCMC options.
}
