% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcProgress.R
\name{mcmcProgressPrint}
\alias{mcmcProgressPrint}
\title{printing progress monitor}
\usage{
mcmcProgressPrint(mcmcloop)
}
\arguments{
\item{mcmcloop}{an mcmc loop iterator}
}
\value{
a progress monitor
}
\description{
a progress monitor that prints each iteration
}
