% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMRF.R
\name{neighLocs}
\alias{neighLocs}
\title{neighLocs function}
\usage{
neighLocs(coord, cellwidth, order)
}
\arguments{
\item{coord}{coordinate of interest}

\item{cellwidth}{a scalar, the width of the grid cells}

\item{order}{the order of the SPDE approximation: see Lindgren et al 2011 for details}
}
\value{
coordinates of centroids of neighbours
}
\description{
A function used in the computation of neighbours on non-rectangular grids. Not intended for general use.
}
\references{
\enumerate{
    \item Benjamin M. Taylor and Barry S. Rowlingson (2017). spatsurv: An R Package for Bayesian Inference with Spatial Survival Models. Journal of Statistical Software, 77(4), 1-32, doi:10.18637/jss.v077.i04.
    \item Finn Lindgren, Havard Rue, Johan Lindstrom. An explicit link between Gaussian fields and Gaussian Markov random fields: the stochastic partial differential equation approach. Journal of the Royal Statistical Society: Series B 73(4)
}
}
