% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMRF.R
\name{neighOrder}
\alias{neighOrder}
\title{neighOrder function}
\usage{
neighOrder(neighlocs)
}
\arguments{
\item{neighlocs}{an object created by the function neighLocs}
}
\value{
the neighbour orders
}
\description{
A function to compute the order of a set of neighbours. Not intended for general use.
}
\references{
\enumerate{
    \item Benjamin M. Taylor and Barry S. Rowlingson (2017). spatsurv: An R Package for Bayesian Inference with Spatial Survival Models. Journal of Statistical Software, 77(4), 1-32, doi:10.18637/jss.v077.i04.
    \item Finn Lindgren, Havard Rue, Johan Lindstrom. An explicit link between Gaussian fields and Gaussian Markov random fields: the stochastic partial differential equation approach. Journal of the Royal Statistical Society: Series B 73(4)
}
}
