\name{NonSpatialLogLikelihood_or_gradient}
\alias{NonSpatialLogLikelihood_or_gradient}
\title{NonSpatialLogLikelihood_or_gradient function}
\usage{
NonSpatialLogLikelihood_or_gradient(surv, X, beta, omega, control,
  loglikelihood, gradient)
}
\arguments{
  \item{surv}{X}

  \item{X}{X}

  \item{beta}{X}

  \item{omega}{X}

  \item{control}{X}

  \item{loglikelihood}{X}

  \item{gradient}{X}
}
\value{
...
}
\description{
A function to
}

