% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocessing.R
\name{posteriorcov}
\alias{posteriorcov}
\title{posteriorcov function}
\usage{
posteriorcov(
  x,
  probs = c(0.025, 0.5, 0.975),
  rmax = NULL,
  n = 100,
  plot = TRUE,
  bw = FALSE,
  corr = FALSE,
  ...
)
}
\arguments{
\item{x}{an object of class mcmcspatsurv}

\item{probs}{vector of probabilities to be fed to quantile function}

\item{rmax}{maximum distance in space to compute this distance up to}

\item{n}{the number of points at which to evaluate the posterior covariance.}

\item{plot}{whether to plot the result}

\item{bw}{Logical. Plot in black/white/greyscale? Default is to produce a colour plot. Useful for producing plots for journals that do not accept colour plots.}

\item{corr}{logical whether to return the correlation function, default is FALSE i.e. returns the covariance function}

\item{...}{other arguments to be passed to matplot function}
}
\value{
produces a plot of the posterior spatial covariance function.
}
\description{
A function to produce a plot of the posterior covariance function with upper and lower quantiles.
}
\seealso{
\link{print.mcmcspatsurv}, \link{quantile.mcmcspatsurv}, \link{summary.mcmcspatsurv}, \link{vcov.mcmcspatsurv},
\link{frailtylag1}, \link{spatialpars}, \link{hazardpars}, \link{fixedpars}, \link{randompars},
\link{baselinehazard}, \link{predict.mcmcspatsurv}, \link{priorposterior}, \link{MCE},
\link{hazardexceedance}
}
