\name{predict.mcmcspatsurv}
\alias{predict.mcmcspatsurv}
\title{predict.mcmcspatsurv function}
\usage{
\method{predict}{mcmcspatsurv}(object, newdata = NULL, type = "density",
  t = NULL, n = 110, probs = c(0.025, 0.5, 0.975), ...)
}
\arguments{
  \item{object}{an object of class mcmcspatsurv}

  \item{newdata}{X}

  \item{type}{can be "density", "hazard" or "survival".
  Default is "density".}

  \item{t}{X}

  \item{n}{X}

  \item{probs}{X}

  \item{...}{other arguments}
}
\value{
...
}
\description{
A function to
}

