% Generated by roxygen2 (4.0.1): do not edit by hand
\name{predict.mcmcspatsurv}
\alias{predict.mcmcspatsurv}
\title{predict.mcmcspatsurv function}
\usage{
\method{predict}{mcmcspatsurv}(object, type = "densityquantile",
  newdata = NULL, t = NULL, n = 110, indx = NULL, probs = c(0.025,
  0.5, 0.975), plot = TRUE, pause = TRUE, ...)
}
\arguments{
\item{object}{an object of class mcmcspatsurv}

\item{type}{can be "densityquantile","density", "hazard" or "survival". Default is "densityquantile".}

\item{newdata}{X}

\item{t}{X}

\item{n}{X}

\item{indx}{X}

\item{probs}{X}

\item{plot}{X}

\item{pause}{X}

\item{...}{other arguments}
}
\value{
...
}
\description{
A function to
}

