% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocessing.R
\name{priorposterior}
\alias{priorposterior}
\title{priorposterior function}
\usage{
priorposterior(
  x,
  breaks = 30,
  ylab = "Density",
  main = "",
  pause = TRUE,
  bw = FALSE,
  ...
)
}
\arguments{
\item{x}{an object inheriting class mcmcspatsurv}

\item{breaks}{see ?hist}

\item{ylab}{optional y label}

\item{main}{optional title}

\item{pause}{logical whether to pause between plots, the default is TRUE}

\item{bw}{Logical. Plot in black/white/greyscale? Default is to produce a colour plot. Useful for producing plots for journals that do not accept colour plots.}

\item{...}{other arguments passed to the hist function}
}
\value{
plots of the prior (red line) and posterior (histogram).
}
\description{
A function to produce plots of the prior (which shows as a red line) and posterior (showing as a histogram)
}
\seealso{
\link{print.mcmcspatsurv}, \link{quantile.mcmcspatsurv}, \link{summary.mcmcspatsurv}, \link{vcov.mcmcspatsurv},
\link{frailtylag1}, \link{spatialpars}, \link{hazardpars}, \link{fixedpars}, \link{randompars},
\link{baselinehazard}, \link{predict.mcmcspatsurv}, \link{posteriorcov}, \link{MCE},
\link{hazardexceedance}
}
