% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocessing.R
\name{randompars}
\alias{randompars}
\title{randompars function}
\usage{
randompars(x)
}
\arguments{
\item{x}{an object of class mcmcspatsurv}
}
\value{
the Y mcmc chains
}
\description{
A function to return the mcmc chains for the spatially correlated frailties
}
\seealso{
\link{print.mcmcspatsurv}, \link{quantile.mcmcspatsurv}, \link{summary.mcmcspatsurv}, \link{vcov.mcmcspatsurv},
\link{frailtylag1}, \link{spatialpars}, \link{hazardpars}, \link{fixedpars},
\link{baselinehazard}, \link{predict.mcmcspatsurv}, \link{priorposterior}, \link{posteriorcov}, \link{MCE},
\link{hazardexceedance}
}
