% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatsurvMisc.R
\name{setupHazard}
\alias{setupHazard}
\title{setupHazard function}
\usage{
setupHazard(dist, pars, grad = FALSE, hess = FALSE)
}
\arguments{
\item{dist}{an object of class 'basehazardspec'}

\item{pars}{parameters with which to create the functions necessary to evaluate the log posterior}

\item{grad}{logical, whetether to create gradient functions for the baseline hazard and cumulative hazard}

\item{hess}{logical, whetether to create hessian functions for the baseline hazard and cumulative hazard}
}
\value{
a list of functions used in evaluating the log posterior
}
\description{
A function to set up the baseline hazard, cumulative hazard and derivative functions for use in evaluating the log posterior.
This fucntion is not intended for general use.
}
