% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatgrid.R
\name{showGrid}
\alias{showGrid}
\title{showGrid function}
\usage{
showGrid(dat, cellwidth, ext = 2, boundingbox = NULL)
}
\arguments{
\item{dat}{any spatial data object whose bounding box can be computed using the function bbox.}

\item{cellwidth}{an initial suggested cellwidth}

\item{ext}{the extension parameter for the FFT transform, set to 2 by default}

\item{boundingbox}{optional bounding box over which to construct computational grid, supplied as an object on which the function 'bbox' returns the bounding box}
}
\value{
a plot showing the grid and the data. Ideally the data should only just fit inside the grid.
}
\description{
A function to show the grid that will be used for a given cellwidth
}
