% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocessing.R
\name{spatialpars}
\alias{spatialpars}
\title{spatialpars function}
\usage{
spatialpars(x)
}
\arguments{
\item{x}{an object of class mcmcspatsurv}
}
\value{
the eta mcmc chains
}
\description{
A function to return the mcmc chains for the spatial covariance function parameters
}
\seealso{
\link{print.mcmcspatsurv}, \link{quantile.mcmcspatsurv}, \link{summary.mcmcspatsurv}, \link{vcov.mcmcspatsurv},
\link{frailtylag1}, \link{hazardpars}, \link{fixedpars}, \link{randompars},
\link{baselinehazard}, \link{predict.mcmcspatsurv}, \link{priorposterior}, \link{posteriorcov}, \link{MCE},
\link{hazardexceedance}
}
