\docType{package}
\name{spatsurv-package}
\alias{spatsurv-package}
\title{spatsurv}
\format{\preformatted{ logi NA
}}
\usage{
spatsurv
}
\description{
An R package for spatially correlated parametric
proportional hazards survial analysis.
}
\details{
\tabular{ll}{ Package: \tab spatsurv\cr Version: \tab
0.9-6\cr Date: \tab 2014-07-10\cr License: \tab GPL-3 \cr }

section{Dependencies}{ The package \code{spatsurv} depends
upon some other important contributions to CRAN in order to
operate; their uses here are indicated:\cr\cr survival, sp,
iterators, RandomFields, fields, rgl, Matrix, stringr,
mcmcplots. }

section{Citation}{ To cite use of \code{lgcp}, the user may
refer to the following work:\cr\cr spatsurv: an \code{R}
Package for Bayesian Inference with Spatial Survival
Models.\cr Benjamin M. Taylor and Barry S. Rowlingson. \cr
Submitted to The Journal Of Statistical Software. }

references{ X }
}
\author{
Benjamin Taylor, Health and Medicine, Lancaster University,
Barry Rowlingson, Health and Medicine, Lancaster University
}
\keyword{package}

