\docType{package}
\name{spatsurv-package}
\alias{spatsurv-package}
\title{spatsurv}
\format{\preformatted{ logi NA
}}
\usage{
spatsurv
}
\description{
An R package for spatially correlated parametric
proportional hazards survial analysis.
}
\details{
\tabular{ll}{ Package: \tab spatsurv\cr Version: \tab
0.9\cr Date: \tab 2014-08-01\cr License: \tab GPL-3 \cr }

section{Dependencies}{ Fill in here ... }

section{Citation}{ Fill in here ... }

references{ Fill in here ... }
}
\author{
Benjamin Taylor, Health and Medicine, Lancaster University,
Barry Rowlingson, Health and Medicine, Lancaster University
}
\keyword{package}

