% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMRF.R
\name{SPDEprec}
\alias{SPDEprec}
\title{SPDEprec function}
\usage{
SPDEprec(a, ord)
}
\arguments{
\item{a}{parameter a, see Lindgren et al 2011.}

\item{ord}{the order of the SPDE model, see Lindgren et al 2011.}
}
\value{
a function used for creating the precision matrix
}
\description{
A function to used in entering elements into the precision matrix of an SPDE model. Not intended for general use.
}
\references{
\enumerate{
    \item Benjamin M. Taylor and Barry S. Rowlingson (2017). spatsurv: An R Package for Bayesian Inference with Spatial Survival Models. Journal of Statistical Software, 77(4), 1-32, doi:10.18637/jss.v077.i04.
    \item Finn Lindgren, Havard Rue, Johan Lindstrom. An explicit link between Gaussian fields and Gaussian Markov random fields: the stochastic partial differential equation approach. Journal of the Royal Statistical Society: Series B 73(4)
}
}
