% Generated by roxygen2 (4.0.1): do not edit by hand
\name{SpikedExponentialCovFct}
\alias{SpikedExponentialCovFct}
\title{SpikedExponentialCovFct function}
\usage{
SpikedExponentialCovFct()
}
\value{
the spiked exponential covariance function
}
\description{
A function to declare and also evaluate a spiked exponential covariance function. This is an exponential covariance function with a nugget.
}

