% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcIterator.R
\name{summary.mcmc}
\alias{summary.mcmc}
\title{summary.mcmc function}
\usage{
\method{summary}{mcmc}(object, ...)
}
\arguments{
\item{object}{an mcmc iterator}

\item{...}{other args}
}
\description{
summary of an mcmc iterator
print out values of an iterator and reset it. DONT call this
in a loop that uses this iterator - it will reset it. And break.
}
