% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocessing.R
\name{summary.mcmcspatsurv}
\alias{summary.mcmcspatsurv}
\title{summary.mcmcspatsurv function}
\usage{
\method{summary}{mcmcspatsurv}(object, probs = c(0.5, 0.025, 0.975), ...)
}
\arguments{
\item{object}{an object inheriting class mcmcspatsurv}

\item{probs}{vector of quantiles to return}

\item{...}{additional arguments}
}
\value{
summary tables to the console
}
\description{
A function to return summary tables from an MCMC run
}
\seealso{
\link{print.mcmcspatsurv}, \link{quantile.mcmcspatsurv}, \link{vcov.mcmcspatsurv},
\link{frailtylag1}, \link{spatialpars}, \link{hazardpars}, \link{fixedpars}, \link{randompars},
\link{baselinehazard}, \link{predict.mcmcspatsurv}, \link{priorposterior}, \link{posteriorcov}, \link{MCE},
\link{hazardexceedance}
}
