% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partial_likelihood.R
\name{timevaryingPL}
\alias{timevaryingPL}
\title{timevaryingPL function}
\usage{
timevaryingPL(
  formula,
  t0,
  t,
  delta,
  dist,
  data,
  ties = "Efron",
  optimcontrol = NULL
)
}
\arguments{
\item{formula}{a formula of the form 'S ~ coef1 + coef2' etc the object S will be created}

\item{t0}{X}

\item{t}{X}

\item{delta}{censoring indicator a vector of 1 for an event and 0 for censoring}

\item{dist}{X}

\item{data}{X}

\item{ties}{X default is Efron}

\item{optimcontrol}{X}
}
\value{
...
}
\description{
A function to
}
