% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatsurvMisc.R
\name{transformweibull}
\alias{transformweibull}
\title{transformweibull function}
\usage{
transformweibull(x)
}
\arguments{
\item{x}{a vector of paramters}
}
\value{
the transformed parameters. For the weibull model, this is the back-transform from 'alpha' and 'lambda' to 'shape' 'scale' (see ?dweibull).
}
\description{
A function to back-transform estimates of the parameters of the weibull baseline hazard function, so they are commensurate with R's inbuilt density functions.
Transforms from (shape, scale) to (alpha, lambda)
}
